+++ front/src/app/users-details/users-details.component.ts
@@ -9,6 +9,8 @@ import {UserService} from '../_services/user.service';
 })
 export class UsersDetailsComponent implements OnInit {
   users: UserModel[];
+  loading: boolean;
+  userDeleted: boolean;
 
   constructor(private userService: UserService) { }
 
@@ -17,19 +19,28 @@ export class UsersDetailsComponent implements OnInit {
   }
 
   getAllUsers() {
+    this.loading = true;
     this.userService.getAllUsers()
       .subscribe(users => {
         this.users = users;
         console.log(users);
+        this.loading = false;
       });
   }
 
 
   deleteUser(id: string) {
-    this.userService.deleteUser(id)
-      .subscribe(() => {
-        this.getAllUsers();
-      });
+    this.loading = true;
+    if (confirm('Delete this user')) {
+      this.userService.deleteUser(id)
+        .subscribe(() => {
+          this.userDeleted = true;
+          this.getAllUsers();
+          setTimeout(() => {
+            this.userDeleted = false;
+          }, 3000);
+        });
+    }
   }
 
 }
